split with js
