+++ src/pages/mytree.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function mytree(){
+    return(
+        <div></div>
+    
+    )
+
+}
