+++ src/styles.css
@@ -13,7 +13,7 @@
 }
 
 h2{
-    color:#634646;
+    color:#E8DBD1;
 }
 a{
     text-decoration: none;
@@ -22,19 +22,23 @@ a{
 button {
     width:100%;
     margin-left:0%;
+    margin-top: 5%;
     padding:.8em;
     font-size: 20px;
     border-radius: 8em;
     border: none;
     color: white;
-    background-color: #ff7981;
-    box-shadow:0 .3em .3em #CD4B53;
+    background-color: #F3A76B;
+    box-shadow:0 .3em .3em #F46A04;
 }
 
 body{
+    background-image: url("../src/img/Background\ \(1\).svg");
+    background-repeat: no-repeat;
+    background-size:cover;
     margin-left: calc(100%/8);
     margin-right: calc(100%/8);
-    background-color: #f7EEE7;
+    background-color: #042439;
 }
 
 
@@ -42,6 +46,7 @@ body{
     margin:2rem;
     font-size: x-large;
     text-align: center;
+    color: #E8DBD1;
 }
 
 .head h2{
@@ -130,6 +135,6 @@ input{
     border-radius: 8em;
     border: none;
     color: white;
-    background-color: #ff7981;
+    background-color: #F3A76B;
     box-shadow:0 .3em .3em #CD4B53;
 }
